# backend/swagger-prod.yaml
openapi: 3.0.0
info:
  title: Channel Connector API
  version: 1.0.0
  description: Channel and FBY SYNC related operations API.
servers:
- url: http://localhost:5000/
  description: BASE_URL  
...
paths:
...
  /shopify/api/get_products_fby/:
    get:
      summary: Step 1) GET_PRODUCTS_FROM_FBY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         - in: query
           name: updated_after
           schema:
             type: string
             example: "2022-06-20 01:01:01"
           required: true
           description: 'Date in format "YYYY-MM-DD HH:mm:ss", Preferably 2 days before current date.'  
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []    
  /shopify/api/get_prices_fby/:
    get:
      summary: Step 2) GET_PRODUCTS_PRICE_FROM_FBY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         - in: query
           name: updated_after
           schema:
             type: string
             example: "2022-06-20 01:01:01"
           required: true
           description: 'Date in format "YYYY-MM-DD HH:mm:ss", Preferably 2 days before current date.'  
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/push_product_shopify/:
    get:
      summary: Step 3) PUSH_PRODUCTS_TO_SHOPIFY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []  
  /shopify/api/push_product_images/:
    get:
      summary: Step 4) PUSH_PRODUCTS_IMAGES_TO_SHOPIFY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
  /shopify/api/push_variants_shopify/:
    get:
      summary: Step 5) (last Step) PUSH_PRODUCTS_VARIANTS_TO_SHOPIFY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []  
  /shopify/api/update_products_shopify/:
    get:
      summary: Step 6) (UPDATE) UPDATE_PRODUCTS_VARIANTS_TO_SHOPIFY
      tags: [Shopify Products and Price Sync]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []  
  /shopify/api/get_shopify_products/:
    get:
      summary: Get Shopify Products and sync location if where its missing. (Step 1 for syncing product)
      tags: [Shopify,Products]
      parameters:
         
         - name: fby_user_id
           in: query
           schema:
             type: string
             example: '1002'
           required: true
           description: ID of the user to get
           
         - name: product_id
           in: query
           schema:
             type: string
             example: '6836919959709'
           required: false
           description: product _id 
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []   
  /shopify/api/get_shopify_location:
    get:
      summary: Get Shopify Location and sync with downloaded products before sending to FBY. Requied only when location id is missing in products table. (Step 2 for syncing product)
      tags: [Shopify,Products]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ChannelId from channel config.
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []       
  /shopify/api/send_products_fby:
    get:
      summary: Send Products to FBY. (Step 3/last for syncing product) 
      tags: [Shopify,Products,Storeden,WooCommerce,Mirakl,Magento]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: '1002'
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
  /shopify/api/get_fby_stock:
    get:
      summary: Get Fby Stock (Step 1 for syncing Stock)
      tags: [Shopify,Stocks,Storeden,WooCommerce,Amazon,Mirakl,Prestashop,Magento]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: '1002'
           required: true
           description: ID of the user to get
         - in: query
           name: updated_after
           schema:
             type: string
             example: "2022-01-01 01:01:01"
           required: true
           description: 'Date in format "YYYY-MM-DD HH:mm:ss", Preferably 2 days before current date.'
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
  /shopify/api/push_stock_shopify:
    get:
      summary: Push Shopify Stock (Step 2/last for syncing Stock)
      tags: [Shopify,Stocks]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/get_shopify_orders:
    get:
      summary: Get Shopify Orders (Step 1 for syncing Orders)
      tags: [Shopify,Orders]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/send_orders_fby:
    get:
      summary: Send Order FBY (Step 2 for syncing Orders)
      tags: [Shopify,Orders,Storeden,WooCommerce,Amazon,Mirakl,Prestashop,Magento]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/send_cancelled_orders_fby:
    get:
      summary: Send Caneclled Orders FBY (Step 3/last for syncing Orders)
      tags: [Shopify,Orders,Storeden,WooCommerce,Amazon,Mirakl,Prestashop,Magento]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/get_track_number: 
    get:
      summary: Get Track Number (Step 1 for syncing Tracking Number)
      tags: [Shopify,Trackings,Storeden,WooCommerce,Amazon,Mirakl,Prestashop,Magento]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /shopify/api/push_tracks_shopify:
    get:
      summary: Push Shopify Tracks (Step 2/last for syncing Tracking Number)
      tags: [Shopify,Trackings]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /api/fby_alert:
    get:
      summary: send Alert to FBY
      tags: [Error]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []     
  /api/get_cron_schedule:
    get:
      summary: get cron schedule for a particular channel or all the channels
      tags: [Master Data]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 0, 39
           description: ID of the channel
         - in: query
           name: cc_operation
           schema:
             type: string
             example: GET_ORDER_FROM_FBY or empty
           description: cc_operation
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.       
  /api/get_logs_by_alert_id:
    get:
      summary: get alert logs based on alertId
      tags: [Master Data]
      parameters:
         - in: query
           name: alertId
           schema:
             type: string
             example: 1234
             required: true
           description: alertId
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.       
  /shopify/api/error_manage/:
    get:
      summary: Error Management API
      tags: [Error]
      parameters:
         - in: query
           name: case
           schema:
             type: string
             enum: [send_Orders_Fby, send_Products_Fby]
             example: send_Orders_Fby
           required: true
           description: Any of [send_Orders_Fby, send_Products_Fby].
         
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []
  /api/fby_apis:
    get:
      summary: FBY master data sync APIs (payment_method_code, cancel_reason, alert_codes, alert_domains, channel_codes, and currency_codes )
      tags: [Master Data]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1002
           required: true
           description: ID of the user to get
         - name: case
           in: query
           schema:
             type: string
             enum: [payment_method_code, cancel_reason, alert_codes, alert_domains, channel_codes, currency_codes]
             example: payment_method_code
             description: Any of [payment_method_code, cancel_reason, alert_codes, alert_domains, channel_codes, currency_codes] master data will be syncd
           required: true
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
  /storeden/api/push_stock_storeden:
    get:
      summary: Push Storeden Stock (1st get_fby_stock then last step for Storeden stock syncing)
      tags: [Storeden,Stocks]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1004
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
  /storeden/api/get_products_list/:
    get:
      summary: Get Storeden Products. (Step 1 for syncing product)
      tags: [Storeden,Products]
      parameters:
         
         - name: fby_user_id
           in: query
           schema:
             type: string
             example: '1004'
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []  
  /storeden/api/get_storeden_orders:
    get:
      summary: Get Storeden Orders (Step 1 for syncing Orders)
      tags: [Storeden,Orders]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1004
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: []       
  /storeden/api/push_track:
    get:
      summary: Push Storeden Tracking (Step 2/last for syncing Tracking Number)
      tags: [Storeden,Trackings]
      parameters:
         - in: query
           name: fby_user_id
           schema:
             type: string
             example: 1004
           required: true
           description: ID of the user to get
      responses:
        400:
          description: Bad request or  Authorization token is missing!
        401:
          description: Authorization failed! 
        404:
          description: Data not found!.  
        409:
          description: Resource already exists!
        200:
          description: Request processed successfully.
      security:
       - jwt: [] 
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...